More file relocation stuff
