+++ commands/user.py
@@ -11,10 +11,10 @@ from scripts.errors import NationsException, CancelledException
 import scripts.rendering as rendering
 from scripts.ui import ConfirmView
 
-from game.constants import brand_color
-from game.actions import new_nation, new_region, new_army, new_fleet
+from data.constants import brand_color
+from game.logic.actions import new_nation, new_region, new_army, new_fleet
 
-from game.structures import structure_types
+from game.objs.structures import structure_types
 from world.world import nation_list
 
 logger = logging.getLogger(__name__)
