+++ mappaint.py
@@ -10,7 +10,7 @@ from world.database import init_db, get_db, save_tiles
 from world.load import load
 from scripts.log import log_setup
 
-from game.map import Tile, Terrain
+from game.objs.map import Tile, Terrain
 from world.world import tile_list
 
 # === DISCLAIMER REGARDING AI-GENERATED CONTENT ===
